Don't validate Paypal certificate.
Since the beginning of 2019 Paypal is transitioning to the Digicert root
certificate. Unfortunately their old code still reference the old
certificate which was causing things to fail on staging and eventually
production. Even worse, Paypal's adaptive payments are not supported
anymore for new users, and they even deleted part of the SDK from
Github. So the only option we have for now is to fork the core SDK code
and remove the check for the certificate.